fix #45 - removed disconnected getDisputeByTitle function to test classes only, since they are not used in the core application
